1.5.2.6

+++ HoardFarm/Model/CollectedData.cs
@@ -6,7 +6,7 @@ public class CollectedData
     public string Sender { get; set; }
     public double Runtime { get; set; }
     public bool HoardFound { get; set; }
-    public ushort TerritoryTyp { get; set; }
+    public uint TerritoryTyp { get; set; }
     public bool? HoardCollected { get; set; }
     public double? MoveTime { get; set; }
     public bool SafetyMode { get; set; }
